[feat] Show filenames of dependencies in debugger
Currently, third party dependencies does not show up on the browser
debugger, which makes it difficult to debug errors not related to
the App files. This commit implements a source map for all
dependencies, giving a name for the debugger to display the file.
This commit also enhances the module system by using POST requests
instead of GET requests when trying to retrieve files from the
backend, given that there is a limit on how many chracters can be
sent via GET requests.