[fix] Ensure correctness of the head of server rendered HTML
Currently, when the server is handling multiple concurrent requests
changes to the head of the HTML of whichever routes are being
rendered will bleed into one another. This commit changes the place
where values are retrieved, ensuring that it happens immidiately
after the synchronous render is over. This commit also copies the
React Helmet library right into the Ethical source, in order to
avoid loosing global states that might be contained within
different versions of the library.