[feat] Warm up lazy module dependencies
Currently, lazy module dependencies are only loaded when the user
makes an action on the page that causes the relevant lazy module to
be required. This commit implements a warm up system, where lazy
module loaders can add the module to a warm up queue, which will
begin loading lazy dependencies as soon as it is done processing
the essential ones. This in turn allows for a smoother user
interface, without pontential lags caused by loading modules.