[fix] Ensure module not found errors are thrown in the browser
Currently, if a module renders fine on the Node server, but is not
compatible with the browser do to usage of Node's core modules, the
error is swallowed. This happens when the module fails to be
required and is fetched from the server, the server (rightly)
responds back with an empty array, indicating that the module and
its dependencies are already in the browser, which isn't quite
right. When the root file is retried based on the cached base line
object, which should have never been cached do to errors, it
behaves as if the issue has been resolved, thus not throwing an
error. This commit fixes this behavior by ensuring that cache only
happens after module evaluation and ensures that only files meant
to be lazy loaded are requested from the server. All else throws an
error.