[feat] Ensure faster recovery after fatal errors
Currently, recovery takes around six seconds if the server failed
to restart properly. This commit reduces the wait to three seconds,
and also adds a socket connection that allows the client to restart
if a file is modified within those three seconds. This greatly
increases the responsiveness of the recovery.