[fix] Ensure that files in the distribution folder are renamed
Currently, a copy of the renamed file with its old name remains in
the distribution folder. This commit implements a dev processor that
removes the old copy of the renamed file. This currently not in use
do to the file watcher not emitting renaming events, but will come
in handy in the future.