[fix] Ensure third party library errors are thrown
Currently, a sourcemap is added to any third party libraries for
the sake of file identification in dev tools, but for some reason,
this seems to cause errors to be swallowed. This commit removes
the file identification source map until further investigation.
Also the file identification was causing a bit of a slow down in
the "sources" tab of dev tools, so removing it might actually be a
good thing.