[fix] Prepare repo to publish on NPM
Currently, repo codebase will fail when published on NPM because
some dependencies are being installed as dev dependencies, file
transpiling step is missing, and the project description is lacking.
This commit corrects the dependency issue by placing dependencies
in the non-dev section, adds a pre-publish file transpiling script,
and a better project description.